Fix the release compatibility
Last commit made the tool only compatible with 9.3 and 9.4.

This fix makes sure the tool is available for larger number of PostgreSQL
releases (at least from 8.4, but 8.0 seems good too).

* replace myalloc with pg_malloc
* replace mystrdup with pg_strdup
* add pg_malloc and pg_strdup code for releases that don't have these
  functions available
* use PQconnectdb instead of PQconnectdbParams for releases less than 9.3
  which don't have PQconnecdbParams
* add better comments

There are still work to do on the reports (some reports aren't available with
older releases, and sometimes columns are added/modified).