Use a custom schema
connectDatabase clears the search_path which is a secure thing to do,
but I can't create the function. It was simpler and probably cleaner to
create my own schema and put the function inside, rather than copy/paste
the connectDatabase code, and drop the search_path change.

Per a report from Frédéric Yhuel.